final/requirements.txt
https://github.com/aparrish/bezmerizing/archive/master.zip
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
Flat==0.3.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.3
oauthlib==3.1.0
PySocks==1.7.1
requests==2.23.0
requests-oauthlib==1.3.0
scipy==1.4.1
six==1.14.0
tweepy==3.8.0
urllib3==1.25.9
Werkzeug==1.0.1
gunicorn
python-dotenv
textblob
langdetect
flask-socketio
opensimplex